Only delete original modules directory if serialisation was successful

+++ VRCOSC.Game/Managers/ModuleManager.cs
@@ -88,8 +88,10 @@ public sealed class ModuleManager : IEnumerable<Module>, ICanSerialise
                 legacySerialisation.Deserialise(module);
             }
 
-            storage.DeleteDirectory("modules");
-            Serialise();
+            if (serialisationManager.Serialise())
+            {
+                storage.DeleteDirectory("modules");
+            }
         }
         else
         {
